pyproject.toml
[project]
name = "hub-sheetflow-backend"
version = "0.1.0"
description = "Backend para automação de planilhas"
authors = [
    {name = "Hedris Pereira",email = "hedrisgts@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.122.0,<0.123.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = '^8.3.5'
pytest-asyncio = '^1.0.0'
pytest-cov = "^6.1.1"
ruff = '^0.11.12'
taskipy = "^1.14.1"
ipdb = "^0.13.13"
testcontainers = "^4.10.0"
mypy = "^1.16.0"
pre-commit = "^4.2.0"
pytest-dotenv = "^0.5.2"
bandit = "^1.8.3"
monkeytype = "^23.3.0"
debugpy = "^1.8.14"


[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = [
    '-v', '-s',
    '--strict-config',
    '--strict-markers',
    '--cov=src',
    '--cov-report=term-missing:skip-covered',
    '--cov-report=html:reports/coverage',
    '--cov-fail-under=100',
    '-p no:warnings',
]
filterwarnings = ['error']
testpaths = ['tests']
markers = [
    'need_dot_env: tests that consult data in a dot env file.',
    'unit: unit tests.',
    'integration: integration tests.',
    'e2e: end-to-end tests.'
]
pythonpath = '.'
env_files = [".env.test"]

[tool.coverage.run]
omit = [
    "src/main.py",
    "src/core/config/logging.py",
    "src/**/__init__.py",
]

[tool.ruff]
target-version = "py313"
exclude = ['.git', '__pycache__', '.mypy_cache', '.pytest_cache', 'migrations', 'fake.py']
fix = true
src = ["src", "tests"]

[tool.ruff.lint]
extend-ignore = ["PLW0603", "PLR0913"]
select = [
    'E',  # pycodestyle errors
    'W',  # pycodestyle warnings
    'F',  # pyflakes
    'I',  # isort (organização de imports)
    'PL', # pylint
    'PT', # flake8-pytest
]

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.ruff.lint.per-file-ignores]
'__init__.py' = ['F401']


[tool.mypy]
strict = true
files = ["src"]
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
explicit_package_bases = true

[tool.bandit]
targets = ["src"]

[tool.taskipy.tasks]
type = 'mypy'
lint = 'ruff check .; ruff check . --diff'
check = 'ruff check . --fix'
format = 'ruff format .'
run = 'python -Om src'
pre_test = 'task lint'
test = 'pytest -s -x --cov=src -vv'
post_test = 'coverage html'
pre_test_unit = 'task lint'
test_unit = 'pytest tests -s -x --cov=src -vv'
pre_test_integration = 'task lint'
test_integration = 'pytest .\tests\integration\ -s -x --cov=src -vv'

[alembic]
script_location = "migrations"